requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
statsmodels
httpx
python-dotenv
pandas
numpy
psutil
gunicorn
google-genai
